move HTML parsing codes (to parse HN text) to a new parser module
