+++ hackernews_tui/src/main.rs
@@ -2,6 +2,7 @@
 pub mod client;
 pub mod config;
 pub mod model;
+pub mod parser;
 pub mod prelude;
 pub mod utils;
 pub mod view;
